fix(accordion): align getItemProps merge behavior and test filename
